Move spaceship movement into another script
