+++ crates/lumina_shared/src/player/spaceship/ability.rs
@@ -1,0 +1,9 @@
+use bevy::prelude::*;
+
+pub(super) struct SpaceshipAbilityPlugin;
+
+impl Plugin for SpaceshipAbilityPlugin {
+    fn build(&self, _app: &mut App) {
+        //
+    }
+}
